+++ include/Raytracing/Texture.h
@@ -14,16 +14,16 @@ class ITexture {
 public:
 	virtual ~ITexture() = default;
 
-	virtual Color value(double u, double v, const Point3& p) const = 0;
+	virtual Color value(float u, float v, const Point3& p) const = 0;
 };
 
 class SolidColor : public ITexture {
 public:
 	SolidColor(const Color& c);
 
-	SolidColor(double r, double g, double b);
+	SolidColor(float r, float g, float b);
 
-	Color value(double u, double v, const Point3& p) const override;
+	Color value(float u, float v, const Point3& p) const override;
 
 private:
 	Color color_val;
@@ -31,14 +31,14 @@ private:
 
 class CheckerTexture : public ITexture {
 public:
-	CheckerTexture(double scale, std::shared_ptr<ITexture> even_tex, std::shared_ptr<ITexture> odd_tex);
+	CheckerTexture(float scale, std::shared_ptr<ITexture> even_tex, std::shared_ptr<ITexture> odd_tex);
 
-	CheckerTexture(double scale, const Color& even_color, const Color& odd_color);
+	CheckerTexture(float scale, const Color& even_color, const Color& odd_color);
 
-	Color value(double u, double v, const Point3 &p) const override;
+	Color value(float u, float v, const Point3 &p) const override;
 
 private:
-	double inv_scale;
+	float inv_scale;
 	std::shared_ptr<ITexture> even;
 	std::shared_ptr<ITexture> odd;
 };
@@ -47,7 +47,7 @@ class ImageTexture : public ITexture {
 public:
 	ImageTexture(const std::string& image, const std::string& parent = IMG_INPUT_DIR);
 
-	Color value(double u, double v, const Point3 &p) const override;
+	Color value(float u, float v, const Point3 &p) const override;
 
 private:
 	Image img;
@@ -55,26 +55,26 @@ private:
 
 class NoiseTexture : public ITexture {
 public:
-	NoiseTexture(double frequency, int octave_count, double persistence);
+	NoiseTexture(float frequency, int octave_count, float persistence);
 
-	Color value(double u, double v, const Point3& p) const override;
+	Color value(float u, float v, const Point3& p) const override;
 private:
 	Perlin noise;
-	double frequency;
+	float frequency;
 	int octave_count;
-	double persistence;
+	float persistence;
 };
 
 class TerrainTexture : public ITexture {
 public:
-	TerrainTexture(double frequency, int octave_count, double persistence);
+	TerrainTexture(float frequency, int octave_count, float persistence);
 
-	Color value(double u, double v, const Point3& p) const override;
+	Color value(float u, float v, const Point3& p) const override;
 private:
 	Perlin noise;
-	double frequency;
+	float frequency;
 	int octave_count;
-	double persistence;
+	float persistence;
 };
 
 #endif // RAYTRACING_TEXTURE_H
